SSR indexable content for search and comparison pages
